remove dead code

+++ Source.cpp
@@ -16,10 +16,10 @@ void ntdll_restore(const char *func_name)
   MAP_NATIVE_MODULE("ntdll.dll", &ntdll_mapped);
   const auto hooked_func_adress = (uint8_t *)resolve_jmp(get_proc_address(ntdll, HASHSTR(func_name)), 1);
   const auto hooked_func_size = static_cast<size_t>(GetSizeOfProc(hooked_func_adress, 1));
-  const auto crc_hooked = crc32c::Crc32c(hooked_func_adress, static_cast<unsigned int>(hooked_func_size));
+  const auto crc_hooked = crc32c::Crc32c(hooked_func_adress, hooked_func_size);
   const auto original_func_adress = (uint8_t *)resolve_jmp(get_proc_address(ntdll_mapped, HASHSTR(func_name)), 1);
   const auto original_func_size = static_cast<size_t>(GetSizeOfProc(original_func_adress, 1));
-  const auto crc_original = crc32c::Crc32c(original_func_adress, static_cast<unsigned int>(original_func_size));
+  const auto crc_original = crc32c::Crc32c(original_func_adress, original_func_size);
   // detect hook and restore bytes
   if (crc_original != crc_hooked)
   {
@@ -46,10 +46,10 @@ void kernelbase_restore(const char *func_name)
   MAP_NATIVE_MODULE("kernelbase.dll", &kernelbase_mapped);
   const auto hooked_func_adress = (uint8_t *)resolve_jmp(get_proc_address(kernelbase, HASHSTR(func_name)), 1);
   const auto hooked_func_size = static_cast<size_t>(GetSizeOfProc(hooked_func_adress, 1));
-  const auto crc_hooked = crc32c::Crc32c(hooked_func_adress, static_cast<unsigned int>(hooked_func_size));
+  const auto crc_hooked = crc32c::Crc32c(hooked_func_adress, hooked_func_size);
   const auto original_func_adress = (uint8_t *)resolve_jmp(get_proc_address(kernelbase_mapped, HASHSTR(func_name)), 1);
   const auto original_func_size = static_cast<size_t>(GetSizeOfProc(original_func_adress, 1));
-  const auto crc_original = crc32c::Crc32c(original_func_adress, static_cast<unsigned int>(original_func_size));
+  const auto crc_original = crc32c::Crc32c(original_func_adress, original_func_size);
   // detect hook and restore bytes
   if (crc_original != crc_hooked)
   {
@@ -80,10 +80,10 @@ void user32_restore(const char *func_name)
     MAP_NATIVE_MODULE("win32u.dll", &win32_u_mapped);
     const auto hooked_func_adress = (uint8_t *)resolve_jmp(get_proc_address(win32_u, HASHSTR(func_name)), 1);
     const auto hooked_func_size = static_cast<size_t>(GetSizeOfProc(hooked_func_adress, 1));
-    const auto crc_hooked = crc32c::Crc32c(hooked_func_adress, static_cast<unsigned int>(hooked_func_size));
+    const auto crc_hooked = crc32c::Crc32c(hooked_func_adress, hooked_func_size);
     const auto original_func_adress = (uint8_t *)resolve_jmp(get_proc_address(win32_u_mapped, HASHSTR(func_name)), 1);
     const auto original_func_size = static_cast<size_t>(GetSizeOfProc(original_func_adress, 1));
-    const auto crc_original = crc32c::Crc32c(original_func_adress, static_cast<unsigned int>(original_func_size));
+    const auto crc_original = crc32c::Crc32c(original_func_adress, original_func_size);
     // detect hook and restore bytes
     if (crc_original != crc_hooked)
     {
@@ -111,10 +111,10 @@ void user32_restore(const char *func_name)
     MAP_NATIVE_MODULE("user32.dll", &user32_mapped);
     const auto hooked_func_adress = (uint8_t *)resolve_jmp(get_proc_address(user_32, HASHSTR(func_name)), 1);
     const auto hooked_func_size = static_cast<size_t>(GetSizeOfProc(hooked_func_adress, 1));
-    const auto crc_hooked = crc32c::Crc32c(hooked_func_adress, static_cast<unsigned int>(hooked_func_size));
+    const auto crc_hooked = crc32c::Crc32c(hooked_func_adress, hooked_func_size);
     const auto original_func_adress = (uint8_t *)resolve_jmp(get_proc_address(user32_mapped, HASHSTR(func_name)), 1);
     const auto original_func_size = static_cast<size_t>(GetSizeOfProc(original_func_adress, 1));
-    const auto crc_original = crc32c::Crc32c(original_func_adress, static_cast<unsigned int>(original_func_size));
+    const auto crc_original = crc32c::Crc32c(original_func_adress, original_func_size);
     // detect hook and restore bytes
     if (crc_original != crc_hooked)
     {
